Refactor client token
