+++ types/tokens.ts
@@ -1,0 +1,3 @@
+export const Tokens = {
+  SoundSwitch: new Uint8Array([82, 253, 252, 7, 33, 130, 101, 79, 22, 63, 95, 15, 154, 98, 29, 114])
+}
